requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
celery==5.3.4
flower==2.0.1
eventlet==0.33.3
redis==5.0.1
pymongo==4.6.0
motor==3.3.2
python-dotenv==1.0.0
httpx==0.25.2
openai>=1.6.1
langchain==0.0.340
langchain-openai==0.0.2
qdrant-client==1.15.0
beautifulsoup4==4.12.2
requests==2.31.0
aiofiles==23.2.1
jinja2==3.1.2
trafilatura==1.6.4
